pyproject.toml
[project]
name = "rag-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "langchain-community>=0.3.23",
    "langchain-openai>=0.3.16",
    "langchain-qdrant>=0.2.0",
    "pypdf>=5.5.0",
    "python-dotenv>=1.1.0",
]
